+++ internal/cli/test/command.go
@@ -99,28 +99,10 @@ func lintTarget(path, testSuffix string) ([]string, error) {
 
 //------------------------------------------------------------------------------
 
-// Run executes the test command for a specified path. The path can either be a
-// config file, a config files test definition file, a directory, or the
-// wildcard pattern './...'.
-func Run(path, testSuffix string, lint bool) bool {
-	return RunAll([]string{path}, testSuffix, lint)
-}
-
 // RunAll executes the test command for a slice of paths. The path can either be
 // a config file, a config files test definition file, a directory, or the
 // wildcard pattern './...'.
-func RunAll(paths []string, testSuffix string, lint bool) bool {
-	return runAll(paths, testSuffix, lint, log.Noop(), nil)
-}
-
-// RunAllWithLogger executes the test command for a slice of paths. The path can
-// either be a config file, a config files test definition file, a directory, or
-// the wildcard pattern './...'.
-func RunAllWithLogger(paths []string, testSuffix string, lint bool, logger log.Modular) bool {
-	return runAll(paths, testSuffix, lint, logger, nil)
-}
-
-func runAll(paths []string, testSuffix string, lint bool, logger log.Modular, resourcesPaths []string) bool {
+func RunAll(paths []string, testSuffix string, lint bool, logger log.Modular, resourcesPaths []string) bool {
 	targets, err := GetTestTargets(paths, testSuffix)
 	if err != nil {
 		fmt.Fprintf(os.Stderr, "Failed to obtain test targets: %v\n", err)
@@ -153,7 +135,7 @@ func runAll(paths []string, testSuffix string, lint bool, logger log.Modular, re
 				return false
 			}
 		}
-		if failCases, err = targets[target].execute(target, resourcesPaths, logger); err != nil {
+		if failCases, err = targets[target].Execute(target, resourcesPaths, logger); err != nil {
 			fmt.Fprintf(os.Stderr, "Failed to execute test target '%v': %v\n", target, err)
 			return false
 		}